+++ baybe/objectives/pareto.py
@@ -1,44 +1,33 @@
-"""Functionality for single-target objectives."""
+"""Functionality for multi-target objectives."""
 
-import gc
 import warnings
 
 import pandas as pd
 from attrs import define, field
-from attrs.validators import instance_of
+from attrs.validators import deep_iterable, instance_of, min_len
 from typing_extensions import override
 
 from baybe.objectives.base import Objective
 from baybe.targets.base import Target
-from baybe.targets.enum import TargetMode
-from baybe.targets.numerical import NumericalTarget
-from baybe.utils.dataframe import get_transform_objects, pretty_print_df
-from baybe.utils.plotting import to_string
+from baybe.utils.basic import to_tuple
+from baybe.utils.dataframe import transform_target_columns
 
 
 @define(frozen=True, slots=False)
-class SingleTargetObjective(Objective):
-    """An objective focusing on a single target."""
+class ParetoObjective(Objective):
+    """An objective handling multiple targets in a Pareto sense."""
 
-    _target: Target = field(validator=instance_of(Target), alias="target")
-    """The single target considered by the objective."""
-
-    @override
-    def __str__(self) -> str:
-        targets_list = [target.summary() for target in self.targets]
-        targets_df = pd.DataFrame(targets_list)
-
-        fields = [
-            to_string("Type", self.__class__.__name__, single_line=True),
-            to_string("Targets", pretty_print_df(targets_df)),
-        ]
-
-        return to_string("Objective", *fields)
+    _targets: tuple[Target, ...] = field(
+        converter=to_tuple,
+        validator=[min_len(2), deep_iterable(member_validator=instance_of(Target))],
+        alias="targets",
+    )
+    "The targets considered by the objective."
 
     @override
     @property
     def targets(self) -> tuple[Target, ...]:
-        return (self._target,)
+        return self._targets
 
     @override
     def transform(
@@ -80,25 +69,6 @@ class SingleTargetObjective(Objective):
                 )
         # <<<<<<<<<< Deprecation
 
-        # Even for a single target, it is convenient to use the existing validation
-        # machinery instead of re-implementing it
-        get_transform_objects(
-            df, [self._target], allow_missing=allow_missing, allow_extra=allow_extra
+        return transform_target_columns(
+            df, self.targets, allow_missing=allow_missing, allow_extra=allow_extra
         )
-
-        target_data = df[self._target.name].copy()
-
-        out = self._target.transform(target_data).to_frame()
-
-        # TODO: Remove hotfix (https://github.com/emdgroup/baybe/issues/460)
-        if (
-            isinstance(t := self._target, NumericalTarget)
-            and t.mode is TargetMode.MIN
-            and t.bounds.is_bounded
-        ):
-            out = -out
-        return out
-
-
-# Collect leftover original slotted classes processed by `attrs.define`
-gc.collect()