+++ baybe/surrogates/__init__.py
@@ -1,6 +1,7 @@
 """BayBE surrogates."""
 
 from baybe.surrogates.bandit import BetaBernoulliMultiArmedBanditSurrogate
+from baybe.surrogates.composite import BroadcastingSurrogate, CompositeSurrogate
 from baybe.surrogates.custom import CustomONNXSurrogate, register_custom_architecture
 from baybe.surrogates.gaussian_process.core import GaussianProcessSurrogate
 from baybe.surrogates.linear import BayesianLinearSurrogate
@@ -12,6 +13,8 @@ __all__ = [
     "register_custom_architecture",
     "BayesianLinearSurrogate",
     "BetaBernoulliMultiArmedBanditSurrogate",
+    "BroadcastingSurrogate",
+    "CompositeSurrogate",
     "CustomONNXSurrogate",
     "GaussianProcessSurrogate",
     "MeanPredictionSurrogate",