+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BCUT2D` encoding for `SubstanceParameter`
 - Stored benchmarking results now include the Python environment and version
 - `qPSTD` acquisition function
+- `ParetoObjective` class for Pareto optimization of multiple targets and corresponding
+  `qLogNoisyExpectedHypervolumeImprovement` acquisition function
+- `BroadcastingSurrogate` class and corresponding `Surrogate.broadcast` method for
+  making single-target surrogate models multi-target compatible
+- `CompositeSurrogate` class for composing multi-target surrogates from single-target
+  surrogates
+- `supports_multi_output` attribute/property to `Surrogate`/`AcquisitionFunction`
 
 ### Changed
 - Acquisition function indicator `is_mc` has been removed in favor of new indicators